Fix implicit conversion warning for simulator
Since messageSize is passed in as a size_t by
sendMessageToNanoappFromHost, craftNanoappMessageFromHost should also
take in a size_t to avoid an implicit conversion warning.

Test: ./run_sim.sh

+++ core/include/chre/core/host_comms_manager.h
@@ -198,7 +198,7 @@ class HostCommsManager : public NonCopyable {
                                                uint16_t hostEndpoint,
                                                uint32_t messageType,
                                                const void *messageData,
-                                               uint32_t messageSize);
+                                               size_t messageSize);
 
   /**
    * Posts a crafted event, craftedMessage, to a nanoapp for processing, and
